polishing

+++ oscar-cbls/src/main/scala/oscar/cbls/search/combinators/Combinators.scala
@@ -1543,8 +1543,8 @@ case class Profile(a:Neighborhood,ignoreInitialObj:Boolean = false) extends Neig
     }
   }
 
-  def gainPerCall:String = if(nbCalls ==0) "NA" else ("" + totalGain / nbCalls)
-  def callDuration:String = if(nbCalls == 0 ) "NA" else ("" + totalTimeSpent / nbCalls)
+  def gainPerCall:String = if(nbCalls ==0) "NA" else ("" + (totalGain / nbCalls).toInt)
+  def callDuration:String = if(nbCalls == 0 ) "NA" else ("" + (totalTimeSpent / nbCalls).toInt)
   //gain in obj/s
   def slope:String = if(totalTimeSpent == 0) "NA" else ("" + (1000 * totalGain.toDouble / totalTimeSpent.toDouble).toInt)
 
@@ -1558,7 +1558,7 @@ case class Profile(a:Neighborhood,ignoreInitialObj:Boolean = false) extends Neig
   def collectThisProfileStatistics:String = (padToLength("" + a,31) + " " +
     padToLength("" + nbCalls,6) + " " +
     padToLength("" + nbFound,6) + " " +
-    padToLength("" + totalGain,8) + " " +
+    padToLength("" + totalGain.toInt,8) + " " +
     padToLength("" + totalTimeSpent,12) + " " +
     padToLength("" + gainPerCall,8) + " " +
     padToLength("" + callDuration,12)+ " " +
